requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-crispy-forms==1.11.0
django-multiselectfield==0.1.12
future==0.18.2
html5lib==1.1
idna==2.10
Pillow==8.1.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.5.59
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
webencodings==0.5.1
xhtml2pdf==0.2.5
